jarvis/requirements.txt
anthropic>=0.40.0
pyaudio>=0.2.14
numpy>=1.26.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
gtts>=2.5.0
playsound>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
pyperclip>=1.8.2
